+++ commands/expiring_certificates_test.go
@@ -7,8 +7,6 @@ import (
 	"regexp"
 	"time"
 
-	"github.com/pivotal-cf/jhanda"
-
 	"github.com/pivotal-cf/om/api"
 	"github.com/pivotal-cf/om/commands"
 	"github.com/pivotal-cf/om/commands/fakes"
@@ -226,18 +224,6 @@ var _ = Describe("ExpiringCertificates", func() {
 
 			err := command.Execute([]string{})
 			Expect(err).To(MatchError(ContainSubstring("could not fetch expiring certificates: an api error")))
-		})
-	})
-
-	Describe("Usage", func() {
-		It("returns usage information for the command", func() {
-			command := commands.NewStagedConfig(nil, nil)
-
-			Expect(command.Usage()).To(Equal(jhanda.Usage{
-				Description:      "This command generates a config from a staged product that can be passed in to om configure-product (Note: credentials are not available and will appear as '***')",
-				ShortDescription: "**EXPERIMENTAL** generates a config from a staged product",
-				Flags:            command.Options,
-			}))
 		})
 	})
 })