+++ commands/staged_config.go
@@ -44,14 +44,6 @@ func NewStagedConfig(service stagedConfigService, logger logger) StagedConfig {
 	}
 }
 
-func (ec StagedConfig) Usage() jhanda.Usage {
-	return jhanda.Usage{
-		Description:      "This command generates a config from a staged product that can be passed in to om configure-product (Note: credentials are not available and will appear as '***')",
-		ShortDescription: "**EXPERIMENTAL** generates a config from a staged product",
-		Flags:            ec.Options,
-	}
-}
-
 func (ec StagedConfig) Execute(args []string) error {
 	if _, err := jhanda.Parse(&ec.Options, args); err != nil {
 		return fmt.Errorf("could not parse staged-config flags: %s", err)
@@ -211,4 +203,12 @@ func (ec StagedConfig) chooseCredentialHandler(productGUID string) configparser.
 	}
 
 	return configparser.NewNilHandler()
+}
+
+func (ec StagedConfig) Usage() jhanda.Usage {
+	return jhanda.Usage{
+		Description:      "This command generates a config from a staged product that can be passed in to om configure-product (Note: credentials are not available and will appear as '***')",
+		ShortDescription: "generates a config from a staged product",
+		Flags:            ec.Options,
+	}
 }