+++ CHANGELOG.md
@@ -47,7 +47,13 @@ can be found in [Pivotal Documentation](docs.pivotal.io/platform-automation).
   This ensures that commands are not kept in `bash` history.
   The environment variable `OM_PASSWORD` will overwrite the password value in `env.yml`.
 
-## 4.2.1 (Unreleased)
+## 4.3.0 - Unreleased
+
+- The same commands you love (`pre-deploy-check`, `staged-config`, and `staged-director-config`)
+  no longer have the EXPERIMENTAL tag.
+  Nothing has changed with them, we literally just forgot to remove these ages ago.
+
+## 4.2.1
 
 ### Bug fixes
 * `interpolate` command now has order precedence when a file or stdin is provided.